+++ src/Presentable.php
@@ -8,8 +8,11 @@ use Illuminate\Database\Eloquent\Model;
 
 abstract class Presentable
 {
-    public function __construct(protected Model $model)
-    {
+    public function __construct(
+        protected Model $model,
+        protected string $attribute,
+        protected mixed $option = null
+    ) {
     }
 
     public function render(): string|null