+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Model classes can now specify presentable classes by passing the attribute name and the class path to the `$presenters` array attribute.
 - If the class cannot be found, a new `InvalidPresentableClass` exception will be thrown.
 - Added the ability to hide presenters from the array by implementing the `IsHidden` interface.
+- Added the presenter options. Options can be passed to presenter classes by passing an array to each item in he `$presenters` array.
+.- Added the ability for presenter classes to have access to the attribute from the `$presenters` array.
 
 ## [1.1.0] 06-07-2021
 ### Changes